Move GitHub types
